[Improve][Core] Refactor common options of column/row
